Rename mutator.ts and orchestrator.ts
