+++ src/index.ts
@@ -11,8 +11,8 @@ export { action, actionCreator } from './actionCreator';
 export { default as applyMiddleware } from './applyMiddleware';
 export { default as createStore } from './createStore';
 export { dispatch } from './dispatcher';
-export { default as mutator } from './mutator';
-import { default as orchestrator } from './orchestrator';
+export { default as mutator } from './mutatorDecorator';
+import { default as orchestrator } from './orchestratorDecorator';
 export { default as getRootStore } from './getRootStore';
 export { mutatorAction, orchestratorAction } from './simpleSubscribers';
 export { useStrict };
