+++ src/simpleSubscribers.ts
@@ -2,8 +2,8 @@ import ActionCreator from './interfaces/ActionCreator';
 import SimpleAction from './interfaces/SimpleAction';
 import Subscriber from './interfaces/Subscriber';
 import { action } from './actionCreator';
-import mutator from './mutator';
-import orchestrator from './orchestrator';
+import mutator from './mutatorDecorator';
+import orchestrator from './orchestratorDecorator';
 
 export function createSimpleSubscriber(decorator: Function) {
     return function simpleSubscriber<T extends SimpleAction>(actionType: string, target: T): T {
